# Conflicts:
#	backend/src/main/scala/com/neo/sk/gypsy/core/RoomActor.scala
#	backend/src/main/scala/com/neo/sk/gypsy/core/RoomManager.scala
#	backend/src/main/scala/com/neo/sk/gypsy/core/UserActor.scala
#	backend/src/main/scala/com/neo/sk/gypsy/gypsyServer/GameServer.scala
#	backend/src/main/scala/com/neo/sk/gypsy/http/AuthService.scala
#	backend/src/main/scala/com/neo/sk/gypsy/http/OutApiService.scala
#	client/src/main/scala/com/neo/sk/gypsy/ClientBoot.scala
#	frontend/src/main/scala/com/neo/sk/gypsy/front/gypsyClient/GameHolder.scala
#	shared/src/main/scala/com/neo/sk/gypsy/shared/ptcl/GypsyGameEvent.scala
#	shared/src/main/scala/com/neo/sk/gypsy/shared/ptcl/WsMsgProtocol.scala

+++ backend/src/main/scala/com/neo/sk/gypsy/http/SessionBase.scala
@@ -67,7 +67,7 @@ trait SessionBase extends SessionSupport with ServiceUtils with CirceSupport{
             Some(GypsySession(
               BaseUserInfo(
                 sessionMap(SessionKeys.userType),
-                sessionMap(SessionKeys.userId).toLong,
+                sessionMap(SessionKeys.userId),
                 sessionMap(SessionKeys.name),
                 sessionMap(SessionKeys.headImg)
               ),
